Rope nodes, lifting and folding ship
`rope_node`, `rope_lift` and `rope_from_roots`, generated into all three -- the
text mirror of `bytenode` and `bytefold`, and portable for the same reason:
the `(base, n)` convergence in the commit before this one.

WHAT A ROPE CARRIES THAT A BYTE TREE DOES NOT is the code-point count and the
ASCII bit, packed into one slot -- the count shifted up by one, the bit at the
bottom. A node is ASCII only when EVERY child is, so the fold is an `and`
rather than a sum, and the probe builds one leaf holding 4 bytes of 2 code
points to watch the bit go false and the two counts stay independent.

TWO VOCABULARY BUGS THAT NO GATE COULD HAVE SEEN, both found by reading the
generated text before running it:

`to-i32` on a `Bool` emitted `((int) ascii)` into Java and C#. Neither language
allows that cast -- Rust's `as` does. It compiled nowhere and `check-names`
cannot see it, because every NAME in it exists; only the shape is wrong. There
is a `bool-bit` form now, and it is a different form from `to-i32` because the
two are different questions in two of the three targets.

`FANOUT` and `FLAT_MAX` pointed at `Bytes`. Rust keeps one copy of each on
`rope` and both ports keep a second copy on `Bytes` beside the one on `Str`, so
the table had to pick a home and had picked the wrong one -- which compiled for
as long as only byte sources used it and broke the first time a rope source
did. They point at `Str` now, which is where Rust's live.

`s_bytes`, `s_count`, `s_ascii` and the empty string stay hand-written and are
reached as siblings: each is O(1) on a rope and has to decode UTF-8 on the
other two tiers, and an empty string is INTERNED rather than allocated, so
unlike `b_empty` the generated half cannot build one.

Gates: bin/test 0, conform 26 rows + 13 GC-stress, 0 FAIL, check-kin green over
35 sources, 130 Rust tests under `--all-targets --features diagnostics`.

+++ runtime/src/strs.rs
@@ -114,7 +114,7 @@ impl Rt {
             n += 1;
             start = end;
         }
-        let v = self.rope_from_roots(base, n);
+        let v = self.rope_from_roots(base, n as u32);
         self.pop_to(base);
         v
     }
